@@ -5,32 +5,35 @@
   "private": true,
   "description": "This is my personal portfolio",
   "license": "MIT",
+  "module": "es6",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
-    "gsap": "^3.5.1",
-    "sass": "^1.49.8",
-    "vue": "^2.6.11",
-    "vue-gtag": "^1.10.0",
+    "@babel/eslint-parser": "^7.18.9",
+    "@fortawesome/fontawesome-svg-core": "^6.1.2",
+    "@fortawesome/free-brands-svg-icons": "^6.1.2",
+    "axios": "^0.27.2",
+    "core-js": "^3.24.1",
+    "gsap": "^3.10.4",
+    "sass": "^1.54.4",
+    "vue": "^3.2.37",
+    "vue-gtag": "^2.0.1",
     "vue-meta": "^2.4.0",
-    "vue-router": "^3.2.0"
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.13",
-    "@vue/cli-plugin-eslint": "^4.5.13",
-    "@vue/cli-plugin-router": "^4.5.13",
-    "@vue/cli-service": "^4.5.13",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-plugin-router": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
     "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.12.0",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.14"
+    "eslint": "^8.21.0",
+    "eslint-plugin-vue": "^9.3.0",
+    "sass-loader": "^13.0.2",
+    "vue-template-compiler": "^2.7.8"
   },
   "eslintConfig": {
     "root": true,
